+++ wp-parsely/src/content-helper/common/content-helper-error.tsx
@@ -25,15 +25,20 @@ export enum ContentHelperErrorCode {
 	ParselyApiResponseContainsError = 'ch_response_contains_error',
 	ParselyApiReturnedNoData = 'ch_parsely_api_returned_no_data',
 	ParselyApiReturnedTooManyResults = 'ch_parsely_api_returned_too_many_results',
-	ParselyApiUnauthorized = 401, // Intentionally without quotes.
-	ParselyInternalServerError = 500, // Intentionally without quotes.
-	ParselySchemaValidationFailed = 422, // Intentionally without quotes.
-	ParselyUpstreamMalformedResponse = 507, // Intentionally without quotes.
-	ParselyUpstreamNotAvailable = 503, // Intentionally without quotes.
 	PluginCredentialsNotSetMessageDetected = 'parsely_credentials_not_set_message_detected',
 	PluginSettingsApiSecretNotSet = 'parsely_api_secret_not_set',
 	PluginSettingsSiteIdNotSet = 'parsely_site_id_not_set',
 	PostIsNotPublished = 'ch_post_not_published',
+
+	// Suggestions API.
+	ParselySuggestionsApiAuthUnavailable = 'AUTH_UNAVAILABLE', // HTTP Code 503.
+	ParselySuggestionsApiNoAuthentication = 'NO_AUTHENTICATION', // HTTP Code 401.
+	ParselySuggestionsApiNoAuthorization = 'NO_AUTHORIZATION', // HTTP Code 403.
+	ParselySuggestionsApiNoData = 'NO_DATA', // HTTP Code 507.
+	ParselySuggestionsApiOpenAiError = 'OPENAI_ERROR', // HTTP Code 500.
+	ParselySuggestionsApiOpenAiSchema = 'OPENAI_SCHEMA', // HTTP Code 507.
+	ParselySuggestionsApiOpenAiUnavailable = 'OPENAI_UNAVAILABLE', // HTTP Code 500.
+	ParselySuggestionsApiSchemaError = 'SCHEMA_ERROR', // HTTP Code 422.
 }
 
 /**
@@ -57,11 +62,21 @@ export class ContentHelperError extends Error {
 			ContentHelperErrorCode.ParselyApiResponseContainsError,
 			ContentHelperErrorCode.ParselyApiReturnedNoData,
 			ContentHelperErrorCode.ParselyApiReturnedTooManyResults,
-			ContentHelperErrorCode.ParselyApiUnauthorized,
 			ContentHelperErrorCode.PluginCredentialsNotSetMessageDetected,
 			ContentHelperErrorCode.PluginSettingsApiSecretNotSet,
 			ContentHelperErrorCode.PluginSettingsSiteIdNotSet,
 			ContentHelperErrorCode.PostIsNotPublished,
+
+			// Don't perform any fetch retries for the Suggestions API due to
+			// its time-consuming operations.
+			ContentHelperErrorCode.ParselySuggestionsApiAuthUnavailable,
+			ContentHelperErrorCode.ParselySuggestionsApiNoAuthentication,
+			ContentHelperErrorCode.ParselySuggestionsApiNoAuthorization,
+			ContentHelperErrorCode.ParselySuggestionsApiNoData,
+			ContentHelperErrorCode.ParselySuggestionsApiOpenAiError,
+			ContentHelperErrorCode.ParselySuggestionsApiOpenAiSchema,
+			ContentHelperErrorCode.ParselySuggestionsApiOpenAiUnavailable,
+			ContentHelperErrorCode.ParselySuggestionsApiSchemaError,
 		];
 
 		this.retryFetch = ! noRetryFetchErrors.includes( this.code );
@@ -70,15 +85,16 @@ export class ContentHelperError extends Error {
 		Object.setPrototypeOf( this, ContentHelperError.prototype );
 
 		// Errors that need rephrasing.
-		if ( this.code === ContentHelperErrorCode.ParselyApiUnauthorized ) {
+		if ( this.code === ContentHelperErrorCode.ParselySuggestionsApiNoAuthorization ) {
 			this.message = __(
 				'This AI-powered feature is opt-in. To gain access, please submit a request ' +
 				'<a href="https://wpvip.com/parsely-content-helper/" target="_blank" rel="noreferrer">here</a>.',
 				'wp-parsely'
 			);
-		} else if ( this.code === ContentHelperErrorCode.ParselyInternalServerError ) {
+		} else if ( this.code === ContentHelperErrorCode.ParselySuggestionsApiOpenAiError ||
+			this.code === ContentHelperErrorCode.ParselySuggestionsApiOpenAiUnavailable ) {
 			this.message = __(
-				'The Parse.ly API returned an internal server error. Please try again later.',
+				'The Parse.ly API returned an internal server error. Please retry with a different input, or try again later.',
 				'wp-parsely'
 			);
 		} else if ( this.code === ContentHelperErrorCode.HttpRequestFailed &&
@@ -87,23 +103,22 @@ export class ContentHelperError extends Error {
 				'The Parse.ly API did not respond in a timely manner. Please try again later.',
 				'wp-parsely'
 			);
-		} else if ( this.code === ContentHelperErrorCode.ParselySchemaValidationFailed ) {
+		} else if ( this.code === ContentHelperErrorCode.ParselySuggestionsApiSchemaError ) {
 			this.message = __(
-				'The Parse.ly API returned a validation error. Please try again later.',
+				'The Parse.ly API returned a validation error. Please try again with different parameters.',
 				'wp-parsely'
 			);
-		} else if ( this.code === ContentHelperErrorCode.ParselyUpstreamMalformedResponse &&
-			this.message.includes( 'Insufficient Storage' ) ) {
+		} else if ( this.code === ContentHelperErrorCode.ParselySuggestionsApiNoData ) {
 			this.message = __(
 				'The Parse.ly API couldn\'t find any relevant data to fulfill the request. Please retry with a different input.',
 				'wp-parsely'
 			);
-		} else if ( this.code === ContentHelperErrorCode.ParselyUpstreamMalformedResponse ) {
+		} else if ( this.code === ContentHelperErrorCode.ParselySuggestionsApiOpenAiSchema ) {
 			this.message = __(
-				'The Parse.ly API returned a malformed response. Please try again later.',
+				'The Parse.ly API returned an incorrect response. Please try again later.',
 				'wp-parsely'
 			);
-		} else if ( this.code === ContentHelperErrorCode.ParselyUpstreamNotAvailable ) {
+		} else if ( this.code === ContentHelperErrorCode.ParselySuggestionsApiAuthUnavailable ) {
 			this.message = __(
 				'The Parse.ly API is currently unavailable. Please try again later.',
 				'wp-parsely'
@@ -136,7 +151,8 @@ export class ContentHelperError extends Error {
 				'wp-parsely'
 			) );
 		}
-		if ( this.code === ContentHelperErrorCode.ParselyApiForbidden ) {
+		if ( this.code === ContentHelperErrorCode.ParselyApiForbidden ||
+			this.code === ContentHelperErrorCode.ParselySuggestionsApiNoAuthentication ) {
 			this.hint = this.Hint( __(
 				"Please ensure that the Site ID and API Secret given in the plugin's settings are correct.",
 				'wp-parsely'