+++ wp-parsely/CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.14.2](https://github.com/Parsely/wp-parsely/compare/3.14.1...3.14.2) - 2024-03-19
+
+### Changed
+
+- PCH Related Posts: Don't show "Filter By" when not needed ([#2299](https://github.com/Parsely/wp-parsely/pull/2299))
+
+### Fixed
+
+- PCH: Resolve cursor jump to start in Paragraph Blocks upon selection ([#2305](https://github.com/Parsely/wp-parsely/pull/2305))
+- PCH: Fix issue with common styles not being correctly bundled ([#2302](https://github.com/Parsely/wp-parsely/pull/2302))
+- PCH Excerpt Generator: Fix missing styling ([#2300](https://github.com/Parsely/wp-parsely/pull/2300))
+- PCH Editor Sidebar: Fix incorrect AI opt-in message ([#2296](https://github.com/Parsely/wp-parsely/pull/2296))
+
+### Dependency Updates
+
+- The list of all dependency updates for this release is available [here](https://github.com/Parsely/wp-parsely/pulls?q=is%3Apr+is%3Amerged+milestone%3A3.14.2+label%3A%22Component%3A+Dependencies%22).
+
 ## [3.14.1](https://github.com/Parsely/wp-parsely/compare/3.14.0...3.14.1) - 2024-03-14
 
 ### Changed